Created This Week's Picks page
Added rating system for recipes. Updated settings.json files so that recipes have default ratings and review count. Updated SignIn page so that the Login failed alert is displayed right under the Card component for login. Package updated for the Rating system. Added new publication so that user does not have to be logged in to view TopPicks page.